# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hotel Price API",
    "description": "\n# HotelPlus Rate Intelligence Platform API\n\nSentinel Core Engine providing:\n- Real-time and scheduled competitor price intelligence.\n- Semantic search using vector embeddings and PostgreSQL pgvector.\n- Otonom reasoning and decision analysis using Gemini models.\n- Deep market yield analysis and rate parity logs.\n\n### Compliance & Security\n- Rate-limited endpoints for system stability.\n- Zero PII footprint in accordance with GDPR & CCPA scope exclusion.\n- All requests are logged and monitored under strict Row-Level Security (RLS).\n",
    "termsOfService": "https://hotelplustr.com/terms",
    "contact": {
      "name": "HotelPlus Compliance and API Support",
      "url": "https://hotelplustr.com/contact",
      "email": "info@hotelplus.com.tr"
    },
    "license": {
      "name": "Proprietary B2B License",
      "url": "https://hotelplustr.com/license"
    },
    "version": "2026.03"
  },
  "paths": {
...
    "/api/hotels": {
      "get": {
        "tags": [
          "Hotels",
          "hotels"
        ],
        "summary": "List Hotels",
        "description": "Retrieves a list of hotels associated with the current user.",
        "operationId": "list_hotels_api_hotels_get",
        "parameters": [
          {
            "name": "include_deleted",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Deleted"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Hotel"
                  },
                  "title": "Response List Hotels Api Hotels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Hotels",
          "hotels"
        ],
        "summary": "Create Hotel",
        "description": "Creates a hotel with plan-based limits, profile self-healing, and\ntoken discovery via the unified service layer.\n\nAGENT_FIX: Migration Logic (March 2026)\nPreviously this route did a raw db.table(\"hotels\").insert(),\nwhich bypassed token discovery and left hotels without property_token/serp_api_id.\nThe cleanup script then deleted them as \"orphans\". Now we delegate to\nadd_hotel_to_account_logic which handles enrichment, directory sync, and logging.",
        "operationId": "create_hotel_api_hotels_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hotel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hotels/{hotel_id}": {
      "patch": {
        "tags": [
          "Hotels",
          "hotels"
        ],
        "summary": "Update Hotel",
        "operationId": "update_hotel_api_hotels__hotel_id__patch",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Hotel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hotel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Hotels",
          "hotels"
        ],
        "summary": "Delete Hotel",
        "operationId": "delete_hotel_api_hotels__hotel_id__delete",
        "parameters": [
          {
            "name": "hotel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Hotel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": [
          "Profile",
          "profile"
        ],
        "summary": "Get Settings",
        "description": "Retrieves user-specific application settings (alert thresholds, notifications).\nIf no settings exist, it initializes them with safe defaults.",
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profile",
          "profile"
        ],
        "summary": "Update Settings",
        "description": "Persists user settings updates.\nHandles both creation (first-time) and modification.",
        "operationId": "update_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/users/{user_id}": {
      "patch": {
        "tags": [
          "Admin",
          "admin"
        ],
        "summary": "Admin Update User",
        "description": "Directly updates a user profile from the admin interface.\nUsed for managing subscriptions, roles, and manual status overrides.",
        "operationId": "admin_update_user_api_admin_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUser"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin",
          "admin"
        ],
        "summary": "Delete Admin User",
        "description": "Deletes a user and their associated data (hotels, scans, logs).\nUsed for compliance and account cleanup.",
        "operationId": "delete_admin_user_api_admin_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin",
          "admin"
        ],
        "summary": "Get Admin Users",
        "description": "Lists all users in the system with their roles and subscription status.\nProvides the core user management view for Tripzy admins.",
        "operationId": "get_admin_users_api_admin_users_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUser"
                  },
                  "title": "Response Get Admin Users Api Admin Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin",
          "admin"
        ],
        "summary": "Create Admin User",
        "description": "Administrative user creation. Used for internal team onboarding.",
        "operationId": "create_admin_user_api_admin_users_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Admin User Api Admin Users Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/logs": {
      "get": {
        "tags": [
          "Admin",
          "admin"
        ],
        "summary": "Get Admin Logs",
        "description": "System activity logs. Audit trail for administrative actions.",
        "operationId": "get_admin_logs_api_admin_logs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminLog"
                  },
                  "title": "Response Get Admin Logs Api Admin Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  },
...
}